feat: moves api error  test
