+++ packages/swapkit/helpers/tsconfig.json
@@ -1,4 +1,9 @@
 {
   "extends": "../../../tools/typescript/base.json",
-  "include": ["src", "./build.ts", "../../../tools/typescript/typings.d.ts"]
+  "include": [
+    "src",
+    "./build.ts",
+    "../../../tools/typescript/typings.d.ts",
+    "../api/src/__tests__/requestClient.test.ts"
+  ]
 }
